+++ vms/rpcchainvm/runtime/container/client_test.go
@@ -1,4 +1,4 @@
-package podman
+package container
 
 import (
 	"bytes"
